requirements.txt
aiohttp~=3.8.3
aiosqlite~=0.17.0
discord.py

requests~=2.28.1
Pillow~=9.3.0